+++ locan/locan_io/utilities.py
@@ -3,15 +3,20 @@
 Utility functions for file input/output.
 
 """
+from __future__ import annotations
+
 import logging
+import re
 from pathlib import Path
 
-__all__ = ["find_pattern_upstream"]
+__all__ = ["find_file_upstream"]
 
 logger = logging.getLogger(__name__)
 
 
-def find_pattern_upstream(sub_directory, pattern, directory=None):
+def find_file_upstream(
+    sub_directory, pattern, regex=None, top_directory=None
+) -> Path | None:
     """
     Search for first upstream parent of sub_directory that contains pattern.
     Return first pattern found.
@@ -21,9 +26,12 @@ def find_pattern_upstream(sub_directory, pattern, directory=None):
     ----------
     sub_directory : str | bytes | os.PathLike
         Directory or file path to start with.
-    pattern : str
-        The same pattern as in :func:`Path.glob`.
-    directory : str | bytes | os.PathLike
+    pattern : str | None
+        glob pattern passed to :func:`Path.glob`
+    regex : str | None
+        regex pattern passed to :func:`re.search` and applied in addition
+        to glob pattern
+    top_directory : str | bytes | os.PathLike
         Directory in which to stop the search.
 
     Returns
@@ -32,12 +40,22 @@ def find_pattern_upstream(sub_directory, pattern, directory=None):
     """
     sub_directory = Path(sub_directory).resolve(strict=True)
 
-    if directory is None:
-        directory = sub_directory.anchor
+    if top_directory is None:
+        top_directory = sub_directory.anchor
+
+    if regex is not None:
+        regex_ = re.compile(regex)
+    else:
+        regex_ = None
 
     for parent in sub_directory.parents:
         file_list = list(parent.glob(pattern))
+
+        if regex is not None:
+            file_list = [
+                file_ for file_ in file_list if regex_.search(str(file_)) is not None
+            ]
         if file_list:
             return file_list[0]
-        if parent == directory:
+        if parent == top_directory:
             return None