+++ CHANGES.rst
@@ -9,10 +9,11 @@ New Features
 ------------
 - Add analysis routine for `SubpixelBias`
 - Add function `merge_metadata` to merge metadata from class instance or file
+- Add class `Files` for managing file selections to be used in batch processing
 
 API Changes
 -----------
-- Add ...
+- Change find_pattern_upstream into find_file_upstream
 
 Bug Fixes
 ---------